requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
Django==2.1.3
djangorestframework==3.9.0
drf-extensions==0.4.0
kombu==4.2.1
passlib==1.7.1
Pillow==5.3.0
pytz==2018.7
six==1.11.0
SQLAlchemy==1.2.13
SQLAlchemy-Utils==0.33.6
vine==1.1.4
